requirements.txt
Django~=3.2.6
django-widget-tweaks~=1.4.8
django-allauth~=0.44.0
Pillow~=8.2.0
django-stdimage~=5.3.0
django-markdownx~=3.0.1
django-mdeditor~=0.1.18
bootstrap4~=0.1.0
django-bootstrap-v5~=1.0.5
Markdown~=3.3.4
django-environ~=0.6.0
gunicorn~=20.1.0